chore(elements): Clean up form components [SDKI-588]
