+++ .changeset/mighty-peas-flash.md
@@ -1,0 +1,5 @@
+---
+"@clerk/elements": patch
+---
+
+Extract common Form components from single file
